+++ tests/testthat/test-survfit2.R
@@ -9,6 +9,15 @@ test_that("survfit2() works", {
   )
 
   expect_identical(
+    survfit2(Surv(time, status) ~ 1, data = survival::lung, cluster = ph.karno) %>%
+      modifyList(val = list(call = NULL, .Environment = NULL)) %>%
+      unclass(),
+    survival::survfit(Surv(time, status) ~ 1, data = survival::lung, cluster = ph.karno) %>%
+      modifyList(val = list(call = NULL)) %>%
+      unclass()
+  )
+
+  expect_identical(
     survfit2(Surv(survival::lung$time, survival::lung$status) ~ 1) %>%
       modifyList(val = list(call = NULL, .Environment = NULL)) %>%
       unclass(),
@@ -16,6 +25,12 @@ test_that("survfit2() works", {
       modifyList(val = list(call = NULL)) %>%
       unclass()
   )
+
+  expect_message(
+    sf <- df_lung %>% survfit2(Surv(time, status) ~ sex, data = .)
+  )
+  expect_true(inherits(sf, "survfit"))
+  expect_false(inherits(sf, "survfit2"))
 
   expect_error(survfit2(formula = mtcars))
   expect_error(survfit2())