+++ R/scale_ggsurvfit.R
@@ -1,14 +1,17 @@
-#' Reduce Scale Padding
+#' Apply Scales
 #'
 #' @description
 #' I have a hard time remembering the syntax to reduce horizontal and vertical
 #' padding on a ggplot2 figure. Hence, a function with default reduced padding.
 #' The default padding arguments in this functions are:
 #' - `scale_x_continuous(expand = c(0.015, 0))`
-#' - `scale_y_continuous(expand = c(0.025, 0))`.
+#' - `scale_y_continuous(expand = c(0.025, 0), limits = c(0, 1), label = scales::label_percent())`.
 #'
 #' If you use this function, you **must** include all scale specifications
 #' that would appear in `scale_x_continuous()` or `scale_y_continuous()`.
+#'
+#' To reset any of the above settings to their ggplot2 default, set the value
+#' to `NULL`, e.g. `y_scales = list(limits = NULL)`
 #'
 #' @param x_scales a named list of arguments that will be passed to `ggplot2::scale_x_continuous()`.
 #' @param y_scales a named list of arguments that will be passed to `ggplot2::scale_y_continuous()`.
@@ -23,17 +26,25 @@
 #'   add_confidence_interval()
 #'
 #' # use the function defaults
-#' ggsurvfit + scale_continuous_reduce_pad()
+#' ggsurvfit + scale_ggsurvfit()
 #'
 #' # specify additional scales
 #' ggsurvfit +
-#'   scale_continuous_reduce_pad(
+#'   scale_ggsurvfit(
 #'     x_scales = list(breaks = 0:9),
 #'     y_scales = list(label = scales::percent, limits = c(0, 1))
 #'   )
-scale_continuous_reduce_pad <- function(x_scales = list(), y_scales = list()) {
+scale_ggsurvfit <- function(x_scales = list(), y_scales = list()) {
   list(
-    rlang::inject(ggplot2::scale_x_continuous(!!!utils::modifyList(x = list(expand = c(0.015, 0)), val = x_scales))),
-    rlang::inject(ggplot2::scale_y_continuous(!!!utils::modifyList(x = list(expand = c(0.025, 0)), val = y_scales)))
+    rlang::inject(ggplot2::scale_x_continuous(
+      !!!utils::modifyList(x = list(expand = c(0.015, 0)),
+                           val = x_scales)
+    )),
+    rlang::inject(ggplot2::scale_y_continuous(
+      !!!utils::modifyList(x = list(expand = c(0.025, 0),
+                                    label = scales::label_percent(),
+                                    limits = c(0, 1)),
+                           val = y_scales)
+    ))
   )
 }