+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Fix in `survfit2()` that allows users to pass arguments with non-standard evaluation, i.e. bare column names. (#90)
 
-* When the `survfit(weights=)` argument is utilized the number at risk, number of observed events, etc. is a non-integer number. The counts in the risk table are now rounded to the nearest integer. (#90)
+* When the `survfit(weights=)` argument is utilized the number at risk, number of observed events, etc. are a non-integer numbers. The counts in the risk table are now rounded to the nearest integer. (#90)
 
 * Adding ability to report comparative p-values from cumulative incidence plots created with `tidycmprsk::cuminc()` (#84)