+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggsurvfit (development version)
 
+* Updated the default behavior of `add_risktable(risktable_group='auto')` to minimize the number of risk tables that appear below the figure. (#117) 
+
 * Added S3 methods `grid.draw.ggsurvfit()` and `grid.draw.ggcuminc()` which in turn allows us to save images from the package directly with `ggplot2::ggsave()` (#107)
 
 * Increased the default font size on the plot and in the risk tables, and added arguments to control font size in the risk table theme. (#103) 
@@ -20,7 +22,7 @@
 
 * Added the `add_quantile(x_value=)` argument that places line segments at the time specified.
 
-* Added the `scale_ggsurvfit()` function that wraps both `ggplot2::scale_x_continuous()` and `ggplot2::scale_y_continuous()` and uses reduced padding (via the `expand=` argument), labels y-axis with percentages (`labels=`), and sets the y-axis limits to `c(0, 1)` (`limits=`). (#82)
+* Added the `scale_ggsurvfit()` function that wraps both `ggplot2::scale_x_continuous()` and `ggplot2::scale_y_continuous()` and uses reduced padding (via the `expand=` argument), labels y-axis with percentages (`labels=`), adds additional break points on the x-axis (`n.breaks=8`), and sets the y-axis limits to `c(0, 1)` (`limits=`). (#82)
 
 * Package now depends on {ggplot2}, meaning that it'll be attached anytime {ggsurvfit} is attached. (#62)